Rebrand to Bulla + add documentation.
