+++ src/main/java/io/github/thatkawaiisam/bulla/ClickHandler.java
@@ -1,0 +1,14 @@
+package io.github.thatkawaiisam.bulla;
+
+import org.bukkit.entity.Player;
+
+public interface ClickHandler {
+
+    /**
+     * On click of item.
+     *
+     * @param player that executed the click.
+     */
+    void click(Player player);
+
+}
