--- src/main/java/io/github/thatkawaiisam/hotbar/ClickHandler.java
@@ -1,9 +1,0 @@
-package io.github.thatkawaiisam.hotbar;
-
-import org.bukkit.entity.Player;
-
-public interface ClickHandler {
-
-    void click(Player player);
-
-}
